Extract AbstractExtensionService
